+++ pyphi/examples.py
@@ -6,12 +6,19 @@
 Example networks and subsystems to go along with the documentation.
 """
 
+# pylint: disable=too-many-lines
+# flake8: noqa
+
+import string
+
 import numpy as np
 
+from .actual import Context
 from .network import Network
 from .subsystem import Subsystem
-from .actual import Context
 from .utils import all_states
+
+LABELS = string.ascii_uppercase
 
 
 def basic_network(cm=False):
@@ -87,7 +94,7 @@ def basic_network(cm=False):
         ])
     else:
         cm = None
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def basic_subsystem():
@@ -98,6 +105,7 @@ def basic_subsystem():
     return Subsystem(net, state, range(net.size))
 
 
+# TODO label nodes
 def residue_network():
     """The network for the residue example.
 
@@ -146,7 +154,7 @@ def residue_network():
     cm[2:4, 0] = 1
     cm[3:, 1] = 1
 
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def residue_subsystem():
@@ -201,7 +209,7 @@ def xor_network():
         [1, 0, 1],
         [1, 1, 0]
     ])
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def xor_subsystem():
@@ -396,19 +404,19 @@ def propagation_delay_network():
 
     for past_state_index, past_state in enumerate(all_states(num_nodes)):
         current_state = [0 for i in range(num_nodes)]
-        if (past_state[2] == 1 or past_state[7] == 1):
+        if past_state[2] == 1 or past_state[7] == 1:
             current_state[0] = 1
-        if (past_state[0] == 1):
+        if past_state[0] == 1:
             current_state[1] = 1
             current_state[8] = 1
-        if (past_state[3] == 1):
+        if past_state[3] == 1:
             current_state[2] = 1
             current_state[4] = 1
-        if (past_state[1] == 1 ^ past_state[5] == 1):
+        if past_state[1] == 1 ^ past_state[5] == 1:
             current_state[3] = 1
-        if (past_state[4] == 1 and past_state[8] == 1):
+        if past_state[4] == 1 and past_state[8] == 1:
             current_state[6] = 1
-        if (past_state[6] == 1):
+        if past_state[6] == 1:
             current_state[5] = 1
             current_state[7] = 1
         tpm[past_state_index, :] = current_state
@@ -423,7 +431,7 @@ def propagation_delay_network():
                    [1, 0, 0, 0, 0, 0, 0, 0, 0],
                    [0, 0, 0, 0, 0, 0, 1, 0, 0]])
 
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def macro_network():
@@ -446,7 +454,7 @@ def macro_network():
                     [1.0, 1.0, 0.3, 0.3],
                     [1.0, 1.0, 0.3, 0.3],
                     [1.0, 1.0, 1.0, 1.0]])
-    return Network(tpm)
+    return Network(tpm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def macro_subsystem():
@@ -506,15 +514,15 @@ def blackbox_network():
 
     for index, past_state in enumerate(all_states(num_nodes)):
         current_state = [0 for i in range(num_nodes)]
-        if (past_state[5] == 1):
+        if past_state[5] == 1:
             current_state[0] = 1
             current_state[1] = 1
-        if (past_state[0] == 1 and past_state[1]):
+        if past_state[0] == 1 and past_state[1]:
             current_state[2] = 1
-        if (past_state[2] == 1):
+        if past_state[2] == 1:
             current_state[3] = 1
             current_state[4] = 1
-        if (past_state[3] == 1 and past_state[4] == 1):
+        if past_state[3] == 1 and past_state[4] == 1:
             current_state[5] = 1
         tpm[index, :] = current_state
 
@@ -527,7 +535,7 @@ def blackbox_network():
         [1, 1, 0, 0, 0, 0]
     ])
 
-    return Network(tpm, cm)
+    return Network(tpm, cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def rule110_network():
@@ -542,7 +550,7 @@ def rule110_network():
                     [1, 1, 1],
                     [1, 1, 1],
                     [0, 0, 0]])
-    return Network(tpm)
+    return Network(tpm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def rule154_network():
@@ -589,7 +597,7 @@ def rule154_network():
         [0, 0, 1, 1, 1],
         [1, 0, 0, 1, 1]
     ])
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def fig1a():
@@ -668,7 +676,7 @@ def fig1a():
         [0, 0, 0, 0, 0, 0],
         [0, 0, 0, 0, 0, 0]
     ])
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def fig3a():
@@ -697,8 +705,7 @@ def fig3a():
         [1, 0, 0, 0],
         [1, 0, 0, 0]
     ])
-    current_state = (1, 0, 0, 0)
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def fig3b():
@@ -727,8 +734,7 @@ def fig3b():
         [1, 0, 0, 0],
         [1, 0, 0, 0]
     ])
-    current_state = (1, 0, 0, 0)
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def fig4():
@@ -763,8 +769,7 @@ def fig4():
         [1, 0, 1],
         [1, 1, 0],
     ])
-    current_state = (1, 0, 0)
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def fig5a():
@@ -798,8 +803,7 @@ def fig5a():
         [1, 0, 1],
         [1, 1, 0]
     ])
-    current_state = (1, 1, 1)
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def fig5b():
@@ -833,8 +837,7 @@ def fig5b():
         [0, 0, 1],
         [0, 1, 0]
     ])
-    current_state = (1, 0, 0)
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 # The networks in figures 4, 6 and 8 are the same.
@@ -985,7 +988,7 @@ def fig16():
         [0, 0, 0, 0, 0, 1, 1],
         [0, 0, 0, 0, 0, 1, 1]
     ])
-    return Network(tpm, connectivity_matrix=cm)
+    return Network(tpm, connectivity_matrix=cm, node_labels=LABELS[:tpm.shape[1]])
 
 ###################################################################
 #                                                                 #
@@ -1011,7 +1014,7 @@ def ac_ex1_network():
         [1, 0, 0],
         [1, 0, 0]
     ])
-    return Network(tpm, cm)
+    return Network(tpm, cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def ac_ex1_context():
@@ -1049,7 +1052,7 @@ def ac_ex2_network():
         [0, 0, 0, 1],
         [0, 0, 0, 1]
     ])
-    return Network(tpm, cm)
+    return Network(tpm, cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def ac_ex2_context():
@@ -1058,7 +1061,6 @@ def ac_ex2_context():
     before_state = (0, 1, 1, 0)
     after_state = (0, 0, 0, 1)
     return Context(net, before_state, after_state, (1,), (3,))
-
 
 
 def ac_ex3_network():
@@ -1079,7 +1081,7 @@ def ac_ex3_network():
         [1, 0, 0],
         [1, 0, 0]
     ])
-    return Network(tpm, cm)
+    return Network(tpm, cm, node_labels=LABELS[:tpm.shape[1]])
 
 
 def ac_ex3_context():