+++ pyphi/subsystem.py
@@ -10,10 +10,9 @@ import itertools
 import numpy as np
 
 from . import cache, config, utils, validate, approximations
-from .constants import (EMD, KLD, L1, ENT, Direction, BIPARTITION,
-                        WEDGE, FULL, LARGEST, SMALLEST)
-from .models import (Bipartition, Concept, Cut, Mice, Mip, Part, Tripartition,
-                     _null_mip)
+from .constants import EMD, ENTROPY_DIFFERENCE, KLD, L1, Direction, LARGEST, SMALLEST
+from .models import (Bipartition, Concept, Cut, KPartition, Mice, Mip, Part,
+                     Tripartition, _null_mip)
 from .network import irreducible_purviews
 from .node import generate_nodes
 
@@ -97,7 +96,8 @@ class Subsystem:
         # have an accesible object-level cache. Just use a simple memoizer
         self._repertoire_cache = repertoire_cache or cache.DictCache()
 
-        self.nodes = generate_nodes(self, labels=True)
+        self.nodes = generate_nodes(self.tpm, self.cm, self.state,
+                                    network.indices2labels(self.node_indices))
 
         validate.subsystem(self)
 
@@ -602,7 +602,16 @@ class Subsystem:
             return _mip(0, None, None)
 
         # Loop over possible MIP partitions
-        partitions = get_partitions(self, mechanism, purview)
+        # TODO: refactor this to a function and share with actual.py
+        # TODO: validate `PARTITION_TYPE` value
+        if config.PARTITION_TYPE == 'BI':
+            partitions = mip_bipartitions(mechanism, purview)
+        elif config.PARTITION_TYPE == 'TRI':
+            partitions = wedge_partitions(mechanism, purview)
+        elif config.PARTITION_TYPE == 'ALL':
+            partitions = all_partitions(mechanism, purview)
+        elif config.PARTITON_TYPE == 'FULL':
+            partitions = approximations.full_cut(mechanism, purview)
 
         for partition in partitions:
             # Find the distance between the unpartitioned and partitioned
@@ -738,7 +747,6 @@ class Subsystem:
                 max_mip = max(mips)
             else:
                 validate.ties(config.TIES)
-
         return Mice(max_mip)
 
     def core_cause(self, mechanism, purviews=False):
@@ -862,7 +870,7 @@ def mip_bipartitions(mechanism, purview):
     numerators = utils.bipartition(mechanism)
     denominators = utils.directed_bipartition(purview)
 
-    for (n, d) in itertools.product(numerators, denominators):
+    for n, d in itertools.product(numerators, denominators):
         if (n[0] or d[0]) and (n[1] or d[1]):
             yield Bipartition(Part(n[0], d[0]), Part(n[1], d[1]))
 
@@ -877,7 +885,7 @@ def wedge_partitions(mechanism, purview):
         -- X - X --
         B    C   D
 
-    See ``pyphi.config.PARTITION_MECHANISMS`` for more information.
+    See ``pyphi.config.PARTITION_TYPE`` for more information.
 
     Args:
         mechanism (tuple[int]): A mechanism.
@@ -924,6 +932,126 @@ def wedge_partitions(mechanism, purview):
             if not compressible(tripart) and tripart not in yielded:
                 yielded.add(tripart)
                 yield tripart
+
+
+def partitions(collection):
+    # all possible partitions
+    # stackoverflow.com/questions/19368375/set-partitions-in-python
+    if len(collection) == 1:
+        yield [collection]
+        return
+
+    first = collection[0]
+    for smaller in partitions(collection[1:]):
+        for n, subset in enumerate(smaller):
+            yield smaller[:n] + [[first] + subset] + smaller[n+1:]
+        yield [[first]] + smaller
+
+
+def k_partitions(collection, k):
+    # Algorithm for generating k-partitions of a collection
+    # codereview.stackexchange.com/questions/1526/finding-all-k-subset-partitions
+    def visit(n, a):
+        ps = [[] for i in range(k)]
+        for j in range(n):
+            ps[a[j + 1]].append(collection[j])
+        return ps
+
+    def f(mu, nu, sigma, n, a):
+        if mu == 2:
+            yield visit(n, a)
+        else:
+            for v in f(mu - 1, nu - 1, (mu + sigma) % 2, n, a):
+                yield v
+        if nu == mu + 1:
+            a[mu] = mu - 1
+            yield visit(n, a)
+            while a[nu] > 0:
+                a[nu] = a[nu] - 1
+                yield visit(n, a)
+        elif nu > mu + 1:
+            if (mu + sigma) % 2 == 1:
+                a[nu - 1] = mu - 1
+            else:
+                a[mu] = mu - 1
+            if (a[nu] + sigma) % 2 == 1:
+                for v in b(mu, nu - 1, 0, n, a):
+                    yield v
+            else:
+                for v in f(mu, nu - 1, 0, n, a):
+                    yield v
+            while a[nu] > 0:
+                a[nu] = a[nu] - 1
+                if (a[nu] + sigma) % 2 == 1:
+                    for v in b(mu, nu - 1, 0, n, a):
+                        yield v
+                else:
+                    for v in f(mu, nu - 1, 0, n, a):
+                        yield v
+
+    def b(mu, nu, sigma, n, a):
+        if nu == mu + 1:
+            while a[nu] < mu - 1:
+                yield visit(n, a)
+                a[nu] = a[nu] + 1
+            yield visit(n, a)
+            a[mu] = 0
+        elif nu > mu + 1:
+            if (a[nu] + sigma) % 2 == 1:
+                for v in f(mu, nu - 1, 0, n, a):
+                    yield v
+            else:
+                for v in b(mu, nu - 1, 0, n, a):
+                    yield v
+            while a[nu] < mu - 1:
+                a[nu] = a[nu] + 1
+                if (a[nu] + sigma) % 2 == 1:
+                    for v in f(mu, nu - 1, 0, n, a):
+                        yield v
+                else:
+                    for v in b(mu, nu - a, 0, n, a):
+                        yield v
+            if (mu + sigma) % 2 == 1:
+                a[nu - 1] = 0
+            else:
+                a[mu] = 0
+        if mu == 2:
+            yield visit(n, a)
+        else:
+            for v in b(mu - 1, nu - 1, (mu + sigma) % 2, n, a):
+                yield v
+    if k == 1:
+        return ([[[item for item in collection]]])
+    else:
+        n = len(collection)
+        a = [0] * (n + 1)
+        for j in range(1, k + 1):
+            a[n - k + j] = j - 1
+        return f(k, n, 0, n, a)
+
+
+def all_partitions(m, p):
+    m = list(m)
+    p = list(p)
+    mechanism_partitions = partitions(m)
+    for mechanism_partition in mechanism_partitions:
+        mechanism_partition.append([])
+        n_mechanism_parts = len(mechanism_partition)
+        max_purview_partition = min(len(p), n_mechanism_parts)
+        for n_purview_parts in range(1, max_purview_partition + 1):
+            purview_partitions = k_partitions(p, n_purview_parts)
+            n_empty = n_mechanism_parts - n_purview_parts
+            for purview_partition in purview_partitions:
+                purview_partition = [tuple(_list)
+                                     for _list in purview_partition]
+                # Extend with empty tuples so purview partition has same size
+                # as mechanism purview
+                purview_partition.extend([() for j in range(n_empty)])
+                # Unique permutations to avoid duplicates empties
+                for permutation in set(itertools.permutations(purview_partition)):
+                    yield KPartition(
+                        *(Part(tuple(mechanism_partition[i]), tuple(permutation[i]))
+                          for i in range(n_mechanism_parts)))
 
 
 def effect_emd(d1, d2):
@@ -986,6 +1114,7 @@ def measure(direction, d1, d2):
     Returns:
         float: The distance between ``d1`` and ``d2``, rounded to |PRECISION|.
     """
+
     if config.MEASURE == EMD:
         dist = emd(direction, d1, d2)
 
@@ -995,36 +1124,14 @@ def measure(direction, d1, d2):
     elif config.MEASURE == L1:
         dist = utils.l1(d1, d2)
 
-    elif config.MEASURE == ENT:
-        dist = utils.ent(d1, d2)
+    elif config.MEASURE == ENTROPY_DIFFERENCE:
+        dist = utils.entropy_difference(d1, d2)
 
     else:
         validate.measure(config.MEASURE)
 
+    # TODO do we actually need to round here?
     return round(dist, config.PRECISION)
 
 
-def get_partitions(subsystem, mechanism, purview):
-    """Compute the set of partitions of a mechanism-purview combination, potentially within a subsystem with given connectivity matrix.
 
-    Args:
-        mechanism (tuple(int)): Tuple of mechanism indices.
-        purview (tuple(int)): Tuple of purview indices.
-        subsystem (Subsystem): The subsystem the mechanism is within.
-
-    Returns:
-        list(partition): List of all partitions to be checked.
-    """
-    if config.PARTITIONS == BIPARTITION:
-        partitions = mip_bipartitions(mechanism, purview)
-
-    elif config.PARTITIONS == WEDGE:
-        partitions = wedge_partitions(mechanism, purview)
-
-    elif config.PARTITIONS == FULL:
-        partitions = approximations.full_cut(mechanism, purview)
-
-    else:
-        validate.partitions(config.PARTITIONS)
-
-    return partitions