+++ CHANGELOG.md
@@ -54,8 +54,24 @@ Next Release
 - Added `utils.purview` for computing the purview of a repertoire.
 - Added an `exceptions` module.
 - Added `utils.repertoire_shape`.
-- Added `config.PARTITION_MECHANISMS` to require that MIP bipartitions strictly
-  partition the mechanism.
+- Added `config.PARTITION_TYPE` to control the ways in which φ-partitions are
+  generated.
+- Added more functions to the `convert` module:
+  - `holi2loli` and `loli2holi` convert decimal indices between **HOLI** and
+    **LOLI** formats.
+  - `holi2loli_state_by_state` and `loli2holi_state_by_state` convert between
+    **HOLI** and **LOLI** formats for state-by-state TPMs.
+  - Added short aliases for some functions:
+    - `h2l` is `holi2loli`
+    - `l2h` is `loli2holi`
+    - `l2s` is `loli_index2state`
+    - `h2s` is `holi_index2state`
+    - `s2h` is `state2loli_index`
+    - `s2l` is `state2holi_index`
+    - `h2l_sbs` is `holi2loli_state_by_state`
+    - `l2h_sbs` is `loli2holi_state_by_state`
+    - `sbn2sbs` is `state_by_node2state_by_state`
+    - `sbs2sbn` is `state_by_state2state_by_node`
 
 ### Refactor
 - Existing macro coarse-grain logic to use `MacroSubsystem` and `CoarseGrain`.